+++ runtime/libcn/include/bennet/utils/vector.h
@@ -6,8 +6,6 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include <cn-executable/utils.h>
-
 #ifdef __cplusplus
 extern "C" {
 #endif
