+++ runtime/libcn/include/cn-replicate/shape.h
@@ -3,8 +3,7 @@
 
 #include <stddef.h>
 #include <string.h>
-
-#include <cn-executable/utils.h>
+#include <bennet/internals/fulm_types.h>
 
 #ifdef __cplusplus
 extern "C" {
