+++ runtime/libcn/include/cn-testing/test.h
@@ -9,7 +9,6 @@
 #include <time.h>
 
 #include <bennet/prelude.h>
-#include <cn-executable/utils.h>
 #include <cn-testing/result.h>
 #include <cn-replicate/lines.h>
 #include <cn-replicate/shape.h>
