+++ runtime/libcn/include/dune
@@ -17,6 +17,7 @@
    runtime/include/cn-executable/fulminate_alloc.h)
   (cn-executable/hash_table.h as runtime/include/cn-executable/hash_table.h)
   (cn-executable/utils.h as runtime/include/cn-executable/utils.h)
+  (cn-executable/error.h as runtime/include/cn-executable/error.h)
   (cn-executable/rts_deps.h as runtime/include/cn-executable/rts_deps.h)
   (cn-executable/cerb_types.h as runtime/include/cn-executable/cerb_types.h)
   (cn-executable/stack.h as runtime/include/cn-executable/stack.h)
